support to transparent textures + alpha textures
made some adjustments to the main loop that now allow for transparent
textures to be drawn properly (without drawing on top of the previous
render). note that when using transparent textures, the "invisible"
material should be used as the mesh's default override, otherwise
transparent materials will simply be drawn on top of the base material.